+++ sys/arch/news68k/dev/zs.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: zs.c,v 1.11 2003/01/01 01:54:44 thorpej Exp $	*/
+/*	$NetBSD: zs.c,v 1.12 2003/01/11 16:00:48 tsutsui Exp $	*/
 
 /*-
  * Copyright (c) 1996 The NetBSD Foundation, Inc.
@@ -172,7 +172,7 @@ zs_match(parent, cf, aux)
 		return 0;
 
 	/* XXX no default address */
-	if (ha->ha_address == -1)
+	if (ha->ha_address == (u_int)-1)
 		return 0;
 
 	addr = IIOV(ha->ha_address);
